+++ common-db/src/test/resources/executor_sql/executor2.sql
@@ -1,0 +1,15 @@
+-- ----------------------------
+-- Records of role
+-- ----------------------------
+INSERT INTO `role` VALUES ('4', 'updater_82', 'updater_d_1');
+INSERT INTO `role` VALUES ('5', 'n_7', 'descp_4');
+INSERT INTO `role` VALUES ('6', 'n_11', 'descp_64');
+INSERT INTO `role` VALUES ('7', 'n_98', 'descp_79');
+INSERT INTO `role` VALUES ('8', 'n_21', 'descp_5');
+INSERT INTO `role` VALUES ('9', 'n_96', 'descp_98');
+INSERT INTO `role` VALUES ('10', 'n_96', 'descp_29');
+INSERT INTO `role` VALUES ('11', 'n_42', 'descp_54');
+INSERT INTO `role` VALUES ('12', 'n_12', 'descp_65');
+INSERT INTO `role` VALUES ('18', 'n_13', 'descp_52');
+INSERT INTO `role` VALUES ('19', 'n_80', 'descp_99');
+INSERT INTO `role` VALUES ('20', 'n_34', 'descp_18');