+++ common-db/src/test/resources/executor_sql/executor_include2.sql
@@ -1,0 +1,7 @@
+SET FOREIGN_KEY_CHECKS=0;
+
+//include executor1.sql;
+//include executor2.sql;
+//include executor_include.sql;
+
+SET FOREIGN_KEY_CHECKS=1;